requirements.txt
neo4j
azure-cognitiveservices-vision-computervision
pillow
azure-ai-textanalytics==5.1.0